requirements.txt
Apache-Airflow
apache-airflow
requests
pytest
setuptools
apache-airflow-providers-apache-livy
apache_airflow_microsoft_fabric_plugin